+++ Editor/TemplFunctionsAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Willykc.Templ.Editor
+{
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
+    public sealed class TemplFunctionsAttribute : Attribute
+    {
+    }
+}